+++ cli.go
@@ -3,25 +3,53 @@ package cfft
 import (
 	"context"
 	"fmt"
+	"regexp"
 	"strings"
+	"sync"
 
 	"github.com/alecthomas/kong"
 	"github.com/fatih/color"
 )
 
 type CLI struct {
-	Test    TestCmd    `cmd:"" help:"test function"`
-	Init    InitCmd    `cmd:"" help:"initialize files"`
-	Diff    DiffCmd    `cmd:"" help:"diff function code"`
-	Publish PublishCmd `cmd:"" help:"publish function"`
-	KVS     KVSCmd     `cmd:"" help:"manage key-value store"`
-	Version VersionCmd `cmd:"" help:"show version"`
+	Test    *TestCmd    `cmd:"" help:"test function"`
+	Init    *InitCmd    `cmd:"" help:"initialize files"`
+	Diff    *DiffCmd    `cmd:"" help:"diff function code"`
+	Publish *PublishCmd `cmd:"" help:"publish function"`
+	KVS     *KVSCmd     `cmd:"" help:"manage key-value store"`
+	Render  *RenderCmd  `cmd:"" help:"render function code"`
+	Util    *UtilCmd    `cmd:"" help:"utility commands"`
+	Version *VersionCmd `cmd:"" help:"show version"`
 
 	Config string `short:"c" long:"config" help:"config file" default:"cfft.yaml"`
 }
 
 type TestCmd struct {
-	CreateIfMissing bool `help:"create function if missing" default:"false"`
+	CreateIfMissing bool   `help:"create function if missing" default:"false"`
+	Run             string `help:"regexp to run test case names" default:""`
+
+	runRegex *regexp.Regexp
+	once     sync.Once
+}
+
+func (cmd *TestCmd) Setup() error {
+	var err error
+	cmd.once.Do(func() {
+		if cmd.Run != "" {
+			cmd.runRegex, err = regexp.Compile(cmd.Run)
+			if err != nil {
+				err = fmt.Errorf("failed to compile regexp %s, %w", cmd.Run, err)
+			}
+		}
+	})
+	return err
+}
+
+func (cmd *TestCmd) ShouldRun(name string) bool {
+	if cmd.runRegex == nil {
+		return true
+	}
+	return cmd.runRegex.MatchString(name)
 }
 
 type VersionCmd struct{}
@@ -43,7 +71,7 @@ func RunCLI(ctx context.Context, args []string) error {
 	}
 
 	var config *Config
-	if cmds[0] != "init" {
+	if cmds[0] != "init" && cmds[0] != "util" {
 		config, err = LoadConfig(ctx, cli.Config)
 		if err != nil {
 			return err
@@ -57,6 +85,11 @@ func RunCLI(ctx context.Context, args []string) error {
 }
 
 func (app *CFFT) Dispatch(ctx context.Context, cmds []string, cli *CLI) error {
+	if cmds[0] == "util" {
+		// util commands don't need kvs
+		return app.RunUtil(ctx, cmds[1], cli.Util)
+	}
+
 	if err := app.prepareKVS(ctx, cli.Test.CreateIfMissing); err != nil {
 		return err
 	}
@@ -75,8 +108,12 @@ func (app *CFFT) Dispatch(ctx context.Context, cmds []string, cli *CLI) error {
 		return app.DiffFunction(ctx, cli.Diff)
 	case "publish":
 		return app.PublishFunction(ctx, cli.Publish)
+	case "render":
+		return app.Render(ctx, cli.Render)
 	case "kvs":
 		return app.ManageKVS(ctx, cmds[1], cli.KVS)
+	case "util":
+		return app.RunUtil(ctx, cmds[1], cli.Util)
 	case "version":
 		//
 	default: